// src/app/model/orgPartCustomConfig.ts
export class OrgPartCustomConfig{
    id: Number;
    status: Boolean;
    optOrgId: Number;
    optInternalName: String;
    optFieldType: String;
    optName: String;
    optBusinessType: String;
    optMinLength: String;
    optMaxLength: String;
    optToolTip: String;
    optMandatory: Boolean;
    optDefaultValue: Boolean;
    optDisplayInWeb: Boolean;
    optDisplayInMobile: Boolean;
    optDisplayInList: Boolean;
    optIntegrationLabel: String;
    optDisplayOrder: Number;
    optAttributeType: String;
    optDataType: String;
    optDatasource: String;
    optData: String;
    optDefault: Boolean;
    optDisplayName: String;
    optDisplayStatus: Boolean;
}